Add user argument parsing, prepare for 2 types of execution
This commit prepares the program to handle 2 modes of execution, which
I call indexed parenting and direct parenting.
With indexed parenting, the program sets assets to belong to their
closest indexed parent (this is the closest URL to the actual asset that
is viewable on the web). For example, if there exists
`"http://example.org/css/eg.css"`, but there is no indexed page on the
internet at `"http://example.org/css"` then the closest indexed parent of
this resource is `"http://example.org"`, and the child resource becomes
`"css/eg.css"`. Alternatively, in direct parenting, the program
disregards whether the direct parent of the resource is indexed in the
internet and just declares its immediate child. In the example above,
the script would declare `"http://example.org/css"` as the parent of the
asset `"eg.css"`. The program can currently handle direct parenting
(also by default).

This commit adds several simplifications and some improved abstraction
to some of the classes.